+++ test-project/ignored/ignored.php
@@ -1,0 +1,6 @@
+<?php
+
+function ignoredFunction(): string
+{
+    return 'This function is ignored.';
+}